+++ changelog.md
@@ -12,7 +12,9 @@
   * FdFitter: now raises a `RuntimeError` when a fit has no data or fittable parameters.
   * FdFitter: now raises a `KeyError` when trying to access data that does not exist.
   * FdFitter: now raises a `RuntimeError` when trying to plot a fit with multiple models without selecting a model using angular brackets `[]` first.
-
+  * FdFitter: when inverting a model with `interpolate=True`, Pylake now raises a `ValueError` if the minimum or maximum is not finite.
+  * FdFitter: a `ValueError` is raised when adding incompatible models.
+  
 #### New features
 
 * Added API for notes, i.e. `file.notes` returns a dictionary of notes.